ValidadorIP/requirements.txt
Flask>=3.0.0
Flask-RESTful>=0.3.10
Flask-SQLAlchemy>=3.1.0
python-dateutil>=2.8.0
requests==2.32.5